Adjust asset serve path in production.
This is accessed via assets.dc311rn.com, which uses cloudfront to cache and serve assets faster than this app does.

The result is that the assets should only be loaded from Node once or twice, and the rest of the time CloudFront will serve the assets. Unfortunately, I didn’t get this folder path correct the first time.