Adjust 404 message, allow lookup at 311.dc.gov.
The OpenData export of 311 data hasn’t been updated since Friday morning at 7:33am. I assume that some export worker has crashed — but it’s causing this site to be a little misleading.

This tweaks the 404 message to provide more context, and a quick form to jump over to 311.dc.gov if the user knows the associated email address.